requirements.txt
#  principais 
kivy==2.2.1
kivymd==1.1.1
kivy_deps.sdl2==0.7.0
kivy_deps.glew==0.3.1
kivy_deps.gstreamer==0.3.4

#  banco de dados 
mysql-connector-python==9.0.0
bcrypt==4.2.0

#  variáveis de ambiente 
python-dotenv==1.0.1

#  API OpenAI e rede 
openai==1.43.0
httpx==0.27.0
requests==2.32.3

#  imagens e utilidades 
pillow==10.4.0
pydantic==2.8.2

#  extrair texto de PDF
pymupdf==1.24.9
numpy==2.1.3
typing_extensions==4.12.2
